+++ apps/web/src/design-system/patterns/index.ts
@@ -1,3 +1,6 @@
 export { AppShell, type AppShellProps, type ShellNavItem } from './AppShell'
 export { PageHeader, type PageHeaderProps } from './PageHeader'
 export { DetailHeader, type DetailHeaderProps } from './DetailHeader'
+export { FilterPanel, FilterField, type FilterPanelProps } from './FilterPanel'
+export { BulkActionBar, type BulkActionBarProps } from './BulkActionBar'
+export { EditModeBar, type EditModeBarProps } from './EditModeBar'